Split ShowCreateTableSuite.
